+++ changelog.md
@@ -2,7 +2,8 @@ v0.3.3
 ------
 
 - Added custom default button style (blue buttons) and added white button style.
-- Tweaked appearance of the calculator selection grid.
+- Improved the appearance of the calculator selection grid.
+- Added grid icon for the IPC2221A Track Current calculator.
 
 v0.3.2 
 ------